pyproject.toml
[tool.poetry]
name = "fast-api-starter-template"
version = "0.1.0"
description = "Project FastAPI template with some configurations already made and authentication implemented, to be used as a basis for other projects"
authors = ["Juliana Marques <julianabmarques@live.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.115.6"
uvicorn = "^0.32.1"
alembic = "^1.14.0"
psycopg = "^3.2.3"
sqlmodel = "^0.0.22"
pydantic-settings = "^2.6.1"
passlib = "^1.7.4"
email-validator = "^2.2.0"
pyjwt = "^2.10.1"
python-multipart = "^0.0.19"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
